+++ hbw/app/controllers/concerns/hbw/controller.rb
@@ -24,6 +24,10 @@ module HBW
       params[:entity_class].to_sym
     end
 
+    def initial_variables
+      params[:initial_variables].to_h
+    end
+
     def csrf_token
       session['_csrf_token'] ||= form_authenticity_token
     end